feat: add support for MetaDatasFile
This commit adds support for MetaDatasFile, which is a file containing metadata in JSON format. The MetaDatasFile struct conforms to the FileDocument protocol and can be used to read and write JSON data. Additionally, this commit adds a convenience initializer to the MetaData class that takes a json_MetaData object and creates a new MetaData object with the appropriate properties. Finally, this commit adds a createMetaData() method to the MetaDatasFile struct that can be used to create MetaData objects from the JSON data in the file.

feat(JMdictSettingsView.swift): add support for MetaDatasFile and file pickers for both JMdictFile and MetaDatasFile. Add buttons to create Mots and MetaDatas. Add loading state to show progress.