chore(project.pbxproj): add new files to Xcode project groups
chore(project.pbxproj): add new files to Xcode project and reorganize groups

feat: add new views for editing Japonais, Mot and Sense entities

Add new views for editing Japonais, Mot and Sense entities. These views allow users to add, delete and modify the properties of these entities. The views are implemented using SwiftUI and CoreData.

feat: add new views and view modifiers

Add TraductionEditSheetView, MetaDataPickerSheeView, EditSheet, and ModifiableText view modifiers.

TraductionEditSheetView is a view that allows the user to edit a translation. It takes a Traduction object and an optional array of excluded languages as parameters.

MetaDataPickerSheeView is a view that allows the user to pick a metadata object. It takes a replacedMetaData binding and an optional array of excluded metadata objects as parameters.

EditSheet is a view modifier that presents an edit sheet for a given entity. It takes an isPresented binding and an entity object as parameters.

ModifiableText is a view modifier that allows the user to edit a text field. It takes a text binding, an editState binding, and an equals UUID as parameters.